refactor: move unstable `std/collections` files
