+++ collections/_unstable_binary_search_node_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { assertEquals, assertStrictEquals } from "../../assert/mod.ts";
-import { BinarySearchNode } from "./_binary_search_node.ts";
+import { assertEquals, assertStrictEquals } from "../assert/mod.ts";
+import { BinarySearchNode } from "./_unstable_binary_search_node.ts";
 
 let parent: BinarySearchNode<number>;
 let child: BinarySearchNode<number>;
